requirements.txt
Django==2.1.4
beautifulsoup4==4.6.3
bleach==3.0.2
dj-database-url==0.5.0
gunicorn==19.9.0
psycopg2==2.7.6.1
readability-lxml==0.7
requests==2.21.0
whitenoise==4.1.2
